Add preventDefault to viewer events
This adds `event.preventDefault` and `event.isDefaultPrevented`
methods to the viewer.js event objects if default is preventable.

NOTE: this does not actually cause any default behavior of events to
actually be prevented. This provides a framework for event handlers to
request that the default behavior be prevented, and for code that
fires events to know about it.