Fix ctrl+click on links in IE
In IE, ctrl+click on anchor tags cannot be caught (and default
prevented) by JavaScript. This means that ctrl+click will not be
caught and forwarded to the linkclick event handlers on the viewer.

This is a workaround where we bind a click handler on a child element
instead, and catch and stop the event.